requirements.txt
python-dotenv>=1.2.1
remnawave>=2.4.4
cloudflare>=4.3.1
pyyaml>=6.0.3
aiogram>=3.24.0
fluent.runtime>=0.4.0
